Rm `maybe/utils` and transfer `some_when` to `maybe/core` (#29)
